Organize Codex intelligent routing modules
