+++ src/ucode/intelligent_routing/__init__.py
@@ -1,0 +1,1 @@
+"""Intelligent-routing integrations for supported coding agents."""
