+++ src/ucode/intelligent_routing/codex_hooks.py
@@ -1,0 +1,124 @@
+"""Codex hook configuration for intelligent subagent routing."""
+
+from __future__ import annotations
+
+import shlex
+import subprocess
+
+from ucode.databricks import build_auth_token_argv
+
+ROUTING_HOOK_COMMAND_MARKER = "codex-router-hook"
+
+
+def sync_intelligent_routing_hooks(doc: dict, state: dict, *, enabled: bool) -> None:
+    """Synchronize ucode-managed routing hooks in a Codex config document."""
+    remove_intelligent_routing_hooks(doc)
+    if not enabled:
+        return
+    hooks = doc.setdefault("hooks", {})
+    for event, groups in _routing_hook_groups(state).items():
+        existing = hooks.get(event)
+        if not isinstance(existing, list):
+            existing = []
+        hooks[event] = [*existing, *groups]
+
+
+def remove_intelligent_routing_hooks(doc: dict) -> bool:
+    """Remove only ucode-managed intelligent-routing hooks."""
+    hooks = doc.get("hooks")
+    if not isinstance(hooks, dict):
+        return False
+    changed = False
+    for event in list(hooks):
+        groups = hooks.get(event)
+        if not isinstance(groups, list):
+            continue
+        kept_groups = []
+        for group in groups:
+            if not isinstance(group, dict):
+                kept_groups.append(group)
+                continue
+            handlers = group.get("hooks")
+            if not isinstance(handlers, list):
+                kept_groups.append(group)
+                continue
+            kept_handlers = [
+                handler
+                for handler in handlers
+                if not (
+                    isinstance(handler, dict)
+                    and ROUTING_HOOK_COMMAND_MARKER in str(handler.get("command") or "")
+                )
+            ]
+            if len(kept_handlers) != len(handlers):
+                changed = True
+            if kept_handlers:
+                group["hooks"] = kept_handlers
+                kept_groups.append(group)
+        if kept_groups:
+            hooks[event] = kept_groups
+        else:
+            hooks.pop(event, None)
+    if not hooks:
+        doc.pop("hooks", None)
+    return changed
+
+
+def _routing_hook_groups(state: dict) -> dict[str, list[dict]]:
+    route_argv = _routing_hook_argv(state, "route-subagent")
+    session_argv = _routing_hook_argv(state, "session-start")
+    subagent_argv = _routing_hook_argv(state, "record-subagent")
+    return {
+        "PreToolUse": [
+            {
+                "matcher": "Agent|.*spawn_agent$",
+                "hooks": [_routing_command_hook(route_argv, status="Routing subagent model")],
+            }
+        ],
+        "SessionStart": [
+            {
+                "matcher": "startup|resume|clear",
+                "hooks": [_routing_command_hook(session_argv)],
+            }
+        ],
+        "SubagentStart": [
+            {
+                "hooks": [_routing_command_hook(subagent_argv)],
+            }
+        ],
+    }
+
+
+def _routing_hook_argv(state: dict, event: str) -> list[str]:
+    workspace = str(state.get("workspace") or "")
+    argv = [
+        build_auth_token_argv(workspace, state.get("profile"), use_pat=bool(state.get("use_pat")))[
+            0
+        ],
+        ROUTING_HOOK_COMMAND_MARKER,
+        event,
+    ]
+    if event != "route-subagent":
+        return argv
+    argv += ["--host", workspace]
+    profile = state.get("profile")
+    if isinstance(profile, str) and profile:
+        argv += ["--profile", profile]
+    if state.get("use_pat"):
+        argv.append("--use-pat")
+    for model in state.get("codex_models") or []:
+        if isinstance(model, str) and model:
+            argv += ["--model", model]
+    return argv
+
+
+def _routing_command_hook(argv: list[str], *, status: str | None = None) -> dict:
+    hook = {
+        "type": "command",
+        "command": shlex.join(argv),
+        "command_windows": subprocess.list2cmdline(argv),
+        "timeout": 35,
+    }
+    if status:
+        hook["statusMessage"] = status
+    return hook
